+++ joey/server/__init__.py
@@ -78,9 +78,9 @@ api = Api(app)
 @api.representation('application/json')
 def output_json(data, code, headers=None):
 
-    if isinstance(data, models.Base):
+    if isinstance(data, model.Base):
         data = db.jsonable(data)
-    elif isinstance(data, list) and len(data) and isinstance(data[0], models.Base):
+    elif isinstance(data, list) and len(data) and isinstance(data[0], model.Base):
         data = db.jsonable(data)
 
     resp = make_response(json.dumps(data), code)
@@ -101,10 +101,10 @@ jwt = JWTManager(app)
 @jwt.user_claims_loader
 def user_claims_loader(identity):
     roles = []
-    if isinstance(identity, models.User):
+    if isinstance(identity, model.User):
         type_ = 'user'
         roles = identity.roles.split(',')
-    elif isinstance(identity, models.Node):
+    elif isinstance(identity, model.Node):
         type_ = 'node'
     elif isinstance(identity, dict):
         type_ = 'container'
@@ -121,7 +121,7 @@ def user_claims_loader(identity):
 @jwt.user_identity_loader
 def user_identity_loader(identity):
 
-    if isinstance(identity, models.Authenticatable):
+    if isinstance(identity, model.Authenticatable):
         return identity.id
     if isinstance(identity, dict):
         return identity
@@ -132,7 +132,7 @@ def user_identity_loader(identity):
 @jwt.user_loader_callback_loader
 def user_loader_callback(identity):
     if isinstance(identity, int):
-        return models.Authenticatable.get(identity)
+        return model.Authenticatable.get(identity)
     else:
         return identity
 
@@ -253,9 +253,9 @@ def connect_pty():
         # At this point we're sure that the user/client/whatever 
         # checks out
         user_or_node_id = get_jwt_identity()
-        auth = models.Authenticatable.get(user_or_node_id)
+        auth = model.Authenticatable.get(user_or_node_id)
 
-        if not isinstance(auth, models.User):
+        if not isinstance(auth, model.User):
             log.error("Sorry, but only users can use this websocket")
             return False
 
@@ -328,9 +328,9 @@ def connect_admin():
         # At this point we're sure that the user/client/whatever 
         # checks out
         user_or_node_id = get_jwt_identity()
-        auth = models.Authenticatable.get(user_or_node_id)
+        auth = model.Authenticatable.get(user_or_node_id)
 
-        if not isinstance(auth, models.User):
+        if not isinstance(auth, model.User):
             log.error("Sorry, but only users can use this websocket")
             return False
 
@@ -486,7 +486,7 @@ def run(ctx, *args, **kwargs):
 
     # Actually start the server
     # app.run(*args, **kwargs)
-    nodes, session = models.Node.get(with_session=True)
+    nodes, session = model.Node.get(with_session=True)
     for node in nodes:
         node.status = 'offline'
     session.commit()
